+++ koupleless-runtime/koupleless-test-suite/src/main/java/com/alipay/sofa/koupleless/test/suite/model/CompatibleTestBizConfig.java
@@ -14,14 +14,42 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.test;
+package com.alipay.sofa.koupleless.test.suite.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 /**
  * @author CodeNoobKing
- * @date 2024/2/20
+ * @date 2024/1/15
  */
-public class BootStrapClass {
-    public void bootstrapBase() {
-        System.out.println("Bootstrap base class");
-    }
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CompatibleTestBizConfig {
+    /**
+     * 当前测试的名字。
+     */
+    private String name;
+
+    /**
+     * 初始化的类名。
+     */
+    private String bootstrapClass;
+
+    /**
+     * 测试的类列表。
+     */
+    private List<String> testClasses;
+
+    /**
+     * 需要被模块加载的类列表。
+     */
+    private List<String> loadByBizClassLoaderPatterns;
+
 }
