+++ koupleless-runtime/koupleless-test-suite/src/main/java/com/alipay/sofa/koupleless/test/suite/model/CompatibleTestConfig.java
@@ -14,11 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.test;
+package com.alipay.sofa.koupleless.test.suite.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 /**
  * @author CodeNoobKing
  * @date 2024/2/20
  */
-public class LoadByTestBizClassB {
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CompatibleTestConfig {
+    // testNG, junit5, junit4, etc.
+    private String                        testFramework;
+
+    // mock test biz configs
+    private List<CompatibleTestBizConfig> testBizConfigs;
 }
