+++ koupleless-runtime/koupleless-test-suite/src/main/java/com/alipay/sofa/koupleless/test/suite/mojo/common/CustomJunit5SummaryGeneratingListener.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.mojo.common;
+package com.alipay.sofa.koupleless.test.suite.mojo.common;
 
 import org.apache.maven.plugin.logging.Log;
 import org.junit.platform.engine.TestExecutionResult;
