+++ koupleless-runtime/koupleless-test-suite/src/test/java/com/alipay/sofa/koupleless/test/suite/mock/LoadByBaseClass.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.test;
+package com.alipay.sofa.koupleless.test.suite.mock;
 
 /**
  * @author CodeNoobKing
