+++ koupleless-runtime/koupleless-test-suite/src/test/java/com/alipay/sofa/koupleless/test/suite/mock/TestJunit5Class.java
@@ -1,4 +1,4 @@
-package com.alipay.sofa.koupleless.test;
+package com.alipay.sofa.koupleless.test.suite.mock;
 
 import com.alipay.sofa.ark.container.service.classloader.BizClassLoader;
 import org.junit.jupiter.api.AfterAll;
